Move Combo resolver functionality to ComboUpdater
